pyproject.toml
[project]
name = "samui"
version = "0.1.0"
description = "SAM3 Web UI for image segmentation"
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "httpx>=0.28",
    "pillow>=11.0",
]
